+++ .storybook/addons.js
@@ -1,3 +1,0 @@
-import '@storybook/addon-knobs/register';
-
-