+++ packages/slate-plugins/src/elements/list/queries/isListItem.ts
@@ -1,0 +1,4 @@
+import { Node } from 'slate';
+import { ListType } from '../types';
+
+export const isListItem = (node: Node) => node.type === ListType.LIST_ITEM;