+++ packages/slate-plugins/src/elements/queries/isBlockActive.ts
@@ -1,5 +1,8 @@
 import { Editor } from 'slate';
 
+/**
+ * Does the selection contain a node of this type
+ */
 export const isBlockActive = (editor: Editor, format: string) => {
   const [match] = Editor.nodes(editor, {
     match: n => n.type === format,