+++ packages/slate-plugins/src/elements/withBlock.ts
@@ -3,9 +3,7 @@ import { PARAGRAPH } from './paragraph';
 import { isBlockActive } from './queries';
 import { ToggleBlockEditor } from './types';
 
-export const withBlock = ({
-  unwrapTypes = [],
-}: { unwrapTypes?: string[] } = {}) => <T extends Editor>(editor: T) => {
+export const withBlock = <T extends Editor>(editor: T) => {
   const e = editor as T & ToggleBlockEditor;
 
   e.toggleBlock = (format: string) => {
@@ -14,13 +12,6 @@ export const withBlock = ({
     Transforms.setNodes(e, {
       type: isActive ? PARAGRAPH : format,
     });
-
-    if (unwrapTypes.length) {
-      Transforms.unwrapNodes(editor, {
-        match: n => unwrapTypes.includes(n.type),
-        split: true,
-      });
-    }
   };
 
   return e;