+++ packages/slate-plugins/src/elements/withBreakEmptyReset.ts
@@ -1,5 +1,6 @@
 import { PARAGRAPH } from 'elements/paragraph';
 import { Editor, Transforms } from 'slate';
+import { isBlockTextEmpty } from './queries';
 
 /**
  * On insert break at the start of an empty block in types,
@@ -7,33 +8,31 @@ import { Editor, Transforms } from 'slate';
  */
 export const withBreakEmptyReset = ({
   types,
-  unwrapTypes = [],
+  onUnwrap,
 }: {
   types: string[];
-  unwrapTypes?: string[];
+  onUnwrap?: any;
 }) => <T extends Editor>(editor: T) => {
   const { insertBreak } = editor;
 
   editor.insertBreak = () => {
-    const match = Editor.above(editor, {
+    const currentNodeEntry = Editor.above(editor, {
       match: n => Editor.isBlock(editor, n),
     });
 
-    if (match) {
-      const [matchingNode] = match;
-      if (types.includes(matchingNode.type)) {
-        if (
-          matchingNode.children[matchingNode.children.length - 1].text
-            .length === 0
-        ) {
+    if (currentNodeEntry) {
+      const [currentNode] = currentNodeEntry;
+
+      if (isBlockTextEmpty(currentNode)) {
+        const parent = Editor.above(editor, {
+          match: n => types.includes(n.type),
+        });
+
+        if (parent) {
           Transforms.setNodes(editor, { type: PARAGRAPH });
 
-          if (unwrapTypes.length) {
-            Transforms.unwrapNodes(editor, {
-              match: n => unwrapTypes.includes(n.type),
-              split: true,
-            });
-          }
+          if (onUnwrap) onUnwrap();
+
           return;
         }
       }